share one gguf reader across the model loaders
